+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - switched to the official maven artifacts
+- updated coffee ingredient recipe keys
 
 ## [0.2.0] - 2024-10-24